highscore on screen

+++ core/src/com/dragons/game/view/componentViews/TimerView.java
@@ -26,7 +26,7 @@ public class TimerView implements IView {
     private Integer worldTimer;
     private boolean timeUp; // true when the world timer reaches 0
     private float timeCount;
-    private float scoreCount;
+    private int scoreCount;
 
     //Scene2D widgets
     private Label countdownLabel;
@@ -95,7 +95,7 @@ public class TimerView implements IView {
 
     public boolean isTimeUp() { return timeUp; }
 
-    public float getScoreCount(){
+    public int getScoreCount(){
         return scoreCount;
     }
 }
